requirements_frozen.txt
aiohttp==3.4.4
appdirs==1.4.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
chardet==3.0.4
discord.py==1.0.0a1633+gdec14fa
humanize==0.5.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.7.5
setuptools-scm==3.1.0
six==1.12.0
websockets==6.0
yarl==1.3.0
